Retry torrentapi on no results.
Potentially addresses #16.

The API can return no results without throwing an error, but will return
results on further requests. This adds a delay of 6s plus two requests to
genuine no result requests.

Also persist the last request time for the API in the unlikely event a clean
interpreter is started to make a request faster than the ratelimit time.